Move capture storage primitives to a new 'database' directory.
