+++ src/database/stream.rs
